Charge V2 changes

+++ lib/features/guideHome/helpers/decimal_handler.dart
@@ -1,7 +1,7 @@
 import 'package:decimal/decimal.dart';
 
 BigInt toBase(Decimal amount, int decimals) {
-  Decimal baseUnit = Decimal.fromInt(10).pow(decimals);
+  Decimal baseUnit = Decimal.fromInt(10).pow(decimals).toDecimal();
 
   Decimal inbase = amount * baseUnit;
 
@@ -9,7 +9,7 @@ BigInt toBase(Decimal amount, int decimals) {
 }
 
 Decimal toDecimal(BigInt amount, int decimals) {
-  Decimal baseUnit = Decimal.fromInt(10).pow(decimals);
+  Decimal baseUnit = Decimal.fromInt(10).pow(decimals).toDecimal();
 
   var d = Decimal.parse(amount.toString());
   var result = d / baseUnit;
